requirements.txt
django-allauth
django-allauth[microsoft]
django-allauth-2fa
Pillow
django-star-ratings
oauthlib
requests